requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
beautifulsoup4==4.12.2
requests==2.31.0
langchain==0.0.350
langchain-openai==0.0.2
openai==1.3.7
pydantic==2.5.2
python-dotenv==1.0.0
aiohttp==3.9.1
lxml==4.9.3
pytest==7.4.3
typing-extensions==4.8.0
